pyproject.toml
[project]
name = "medi-cabinet-bot"
version = "0.1.0"
description = "A Telegram bot helping Bangladeshi families reduce medicine waste through smart inventory tracking"
authors = [{name = "Ahnaf Tanjid"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["telegram", "bot", "medicine", "inventory", "health", "bangladesh"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Home Automation",
]

dependencies = [
    "python-telegram-bot[all]>=21.0,<22.0",
    "aiosqlite>=0.19.0,<1.0.0",
    "loguru>=0.7.0,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "fuzzywuzzy>=0.18.0,<1.0.0",
    "python-Levenshtein>=0.25.0,<1.0.0",
    "alembic>=1.13.0,<2.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "python-dateutil>=2.8.0,<3.0.0",
    "httpx>=0.27.0,<1.0.0",
    "Pillow>=10.0.0,<12.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "black>=24.0.0,<25.0.0",
    "ruff>=0.1.0,<1.0.0",
    "mypy>=1.8.0,<2.0.0",
]

[project.urls]
Homepage = "https://github.com/ahnaftanjid/medi-cabinet"
Repository = "https://github.com/ahnaftanjid/medi-cabinet"
Issues = "https://github.com/ahnaftanjid/medi-cabinet/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "config"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "migrations",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true
exclude = [
    "migrations",
    "tests",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=src",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--cov-report=xml",
    "-v",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "tests/*",
    "**/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
